requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
Deprecated==1.2.18
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==25.4.2
geventhttpclient==2.3.3
greenlet==3.2.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.0.0
locust==2.35.0
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
validators==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
